MapRequest: don't fake map/unmap events on recapture
When recapturing windows on restart -- and presumably via restoration of
the window flags written out to the state file -- some clients would end
up in an iconified state, despite having never been in an iconified
state.

It seems to happen intermittently, and can be difficult to reproduce --
but putting fvwm3 on a restart loop can cause this.  It's unlikely to be
a race-condition, despite the symptoms manifesting that way.

Instead, only send fake map/unmap events when recaptured windows are in
the iconified state.